refactor(mcp): split the 3,290-line buildServer into per-tool modules
buildServer was one ~3,290-line function registering all 10 tools inline plus ~18
shared closures and ~20 pure helpers. Split it, with no behavior change:

- mcp-util.ts — the pure module helpers + constants (json, err, doc, clipDoc,
  parseLineRange, summarize*, toBase64, bundleFileChanges, runnerOnline, ...)
- mcp-tool-context.ts — the ToolContext (per-request state + the shared closures:
  resolveWs, reach, wsArg, workQueue, grantedWorkspaces, askableContexts, ...) and
  makeToolContext(base)
- mcp-tools/<name>.ts — one file per tool (list-workspaces, find, read, organize,
  catch-up, comment, stage, publish, checkpoint, use), each a register<Name>Tool(tc);
  tool-local helpers moved in with their tool
- mcp.ts — buildServer now just constructs the server + resources + instructions,
  builds the ToolContext, and calls the 10 register functions IN ORDER (order is
  load-bearing for the surface); 3290 -> 416 lines

Pure reorg: identical tool names, descriptions, schemas, gating, and output. Full gate
green (ci + typecheck + 1143 api tests).
